Changed Damage from float to int, and small animation fix

+++ Assets/Scripts/Combat/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public abstract void TakeDamage(float damage);
+    public abstract void TakeDamage(int damage);
 }
